+++ .gitignore
@@ -6,3 +6,4 @@
 
 # Python cache
 __pycache__/
+logs/
